r/aws_iam_role: Reduce visibility.
